LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scriptnonce-blocked.html

<!DOCTYPE html>
<html>
<head>
<script src='../resources/multiple-iframe-test.js'></script>
<script>
var tests = [
    ['yes', 'script-src 127.0.0.1:8000', 'resources/script.js', 'nonce'],
    ['yes', 'script-src 127.0.0.1:8000 \'nonce-nonce\'', 'resources/script.js', 'nonce'],
    ['yes', 'script-src 127.0.0.1:8000 \'nonce-base64has+and/characters\'', 'resources/script.js', 'base64has+and/characters'],
    ['yes', 'script-src \'nonce-nonce\'', 'resources/script.js', 'notnonce'],
    ['yes', 'script-src \'nonce-notnonce\'', 'resources/script.js', 'nonce'],
];
</script>
</head>
<body onload="testExperimentalPolicy()">
  <p>
    Only the first two of these scripts should execute even though there are parse errors in the policy.

    Update: Since we no longer support X-WebKit-CSP all the scripts should be executed.
  </p>
